% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdb.R
\name{addNet}
\alias{addNet}
\title{Title}
\usage{
addNet(which.one, tissue, netfile, ctfile, gofile, exprdatafile, overwrite)
}
\arguments{
\item{which.one}{The category under which the network is under}

\item{tissue}{The name of the network under the category}

\item{netfile}{The RDS file produced by getDownstreamNetwork()}

\item{ctfile}{The file name of cell type enrichment for the network modules}

\item{gofile}{The Gene Ontology and others enrichment of modules in the network}

\item{exprdatafile}{Expression data used to generate the network}

\item{overwrite}{If TRUE and the network is already added, it overwrites the contents}
}
\description{
Title
}
