% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdb.R
\name{addNetworkToDDBB}
\alias{addNetworkToDDBB}
\title{Title}
\usage{
addNetworkToDDBB(netf, folder, which.one, tissue, rewrite = F,
  filter = c("GO", "KEGG", "REAC", "HP"), ensembl = TRUE, do.go = T,
  do.ct = T, exclude.iea = T, correction.method = "gSCS",
  organism = "hsapiens", out.file = paste0(netf, "_gprofiler.csv"))
}
\arguments{
\item{netf}{The RDS file produced by getDownstreamNetwork()}

\item{folder}{This method will produce enrichment files. This variable
indicates where to store them}

\item{which.one}{Category under which the network should be installed.
If new, the category will be created}

\item{tissue}{A label to designate the network under the category}

\item{rewrite}{If TRUE and the network is already added with that name
under the category, it is overwritten with the new material}

\item{filter}{gProfiler parameter to indicate which annotation databases to
use when assessing for enrichment at the network modules}

\item{ensembl}{Whether the gene names are indicated as Ensembl ids or Symbols}

\item{do.go}{If FALSE the annotation databases enrichment is not perfomed}

\item{do.ct}{If FALSE the cell enrichment is not performed. Note that cell
enrichment works fine for brain. Not so good for other tissues or conditions.}

\item{exclude.iea}{gProfiler parameter to indicate that if FALSE,
inferred annotations from GO are not used}

\item{correction.method}{gProfiler parameter to indicate the p.value adjustment method}

\item{organism}{gProfiler parameter to indicate the organism upon which genes are defined}

\item{out.file}{Optional file to allocate the GO enrichment}
}
\value{
This method updates the global variable coexp.nets with a new entry to it can
be used with the rest of methods of the package
}
\description{
Title
}
