% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{fromEntrez2Ensembl}
\alias{fromEntrez2Ensembl}
\title{Title}
\usage{
fromEntrez2Ensembl(genes, table.file = biotype.all.file,
  ignore.unknown = FALSE, which.are.unknown = FALSE)
}
\arguments{
\item{which.are.unknown}{}
}
\description{
Title
}
