% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{fromGeneName2Ensembl}
\alias{fromGeneName2Ensembl}
\title{Title}
\usage{
fromGeneName2Ensembl(genes, ignore.unknown = FALSE,
  which.are.unknown = FALSE)
}
\arguments{
\item{which.are.unknown}{}
}
\description{
Title
}
