% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{genCrossTabPlot}
\alias{genCrossTabPlot}
\title{Title Module correspondence between two networks}
\usage{
genCrossTabPlot(
  colors1,
  colors2,
  tissue1 = "Net 1",
  tissue2 = "Net 2",
  plot.file = NULL
)
}
\arguments{
\item{colors1}{The `net$moduleColors` element of the 1st network}

\item{colors2}{The `net$moduleColors` element of the 2nd network}

\item{tissue1}{Tissue label as used in the plot for the 1st network}

\item{tissue2}{Tissue label as used in the plot for the 2nd network}

\item{plot.file}{Convenient for generating a pdf when the number of modules is high
so everything fits in the plot nicely.}
}
\value{
P-value matrix of the overlap between net 1 modules (rows) and net 2 modules (columns)
}
\description{
This method is useful to compare two networks created from the same gene pool but
with different method or different expression data source. Note that genes must be
in the same order within the two networks `moduleColors` variable
}
