% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateBetaStudy}
\alias{generateBetaStudy}
\title{Title}
\usage{
generateBetaStudy(
  expr.data,
  powers = c(1:30),
  title = NULL,
  plot.file = NULL,
  net.type = "signed",
  cor.type = "pearson"
)
}
\arguments{
\item{cor.type}{}
}
\value{

}
\description{
Title
}
