% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getAndPlotNetworkLong}
\alias{getAndPlotNetworkLong}
\title{Title Create a WGCNA network + TOM + some plots
If you are not interested in the k-means or before getting into that you want to
have a look at the basic WGCNA network use this method}
\usage{
getAndPlotNetworkLong(
  expr.data,
  beta,
  net.type = "signed",
  tissue.name = "MyTissue",
  title = NULL,
  additional.prefix = NULL,
  min.cluster.size = 100,
  save.plots = TRUE,
  return.tom = FALSE,
  excludeGrey = FALSE,
  max.k.cutoff = 150,
  r.sq.cutoff = 0.8,
  cor.type = "pearson",
  silent = T
)
}
\arguments{
\item{expr.data}{Aan object or a file, genes at columns, samples at rows.
The network net$moduleColors vector will be generated as genes are in the data.frame.}

\item{beta}{If beta is < 0 then it is obtained automatically. You can set your own.}

\item{net.type}{Either "signed" or "unsigned" for easier or trickier MM values interpretation}

\item{tissue.name}{The label you want the code to use for naming the network}

\item{title}{For possible plots}

\item{additional.prefix}{Something you want to add to the naming of files generated to avoid
potential clashes}

\item{min.cluster.size}{Minimum number of genes for a group of them to be a cluster}

\item{save.plots}{You want plots?}

\item{return.tom}{Set this to TRUE if you want the TOM returned with the rest of stuff (it may be really big)}

\item{excludeGrey}{Discard grey genes from the result network}

\item{max.k.cutoff}{Connectivity parameter}

\item{r.sq.cutoff}{Only beta values above this value for the adjusted linear regression model are
considered}

\item{cor.type}{Values in "pearson", "kendall", "spearman"}
}
\value{
If all goes well, a network with the following elements (as a list)
* MEs will be a matrix with the eigengenes (columns) and the values for each sample (rows)
* moduleLabels A vector of labels corresponding to into which cluster each gene is.
labels appear in the same order as genes were disposed in the expression data matrix
* moduleColors A vector of colors corresponding to into which cluster each gene is. Useful to
refer to each module with a color and to signal them at plots. The vector is named with the
genes and they appear in the same order as genes were disposed in the expression data matrix
beta Is the beta value used
type The type of network as in "signed" or "unsigned"
geneTree The dendrogram from which the partition was generated
}
\description{
Title Create a WGCNA network + TOM + some plots
If you are not interested in the k-means or before getting into that you want to
have a look at the basic WGCNA network use this method
}
