% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdb.R
\name{getCellTypeFromTissue}
\alias{getCellTypeFromTissue}
\title{Title}
\usage{
getCellTypeFromTissue(tissue = "SNIG", which.one = "rnaseq", module = NULL)
}
\arguments{
\item{module}{}
}
\value{

}
\description{
Title
}
