% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getMM}
\alias{getMM}
\title{Title Obtain the Module Membership of genes
This function works by getting the Pearson correlation between the
eigengene of the module and the gene(s).}
\usage{
getMM(
  net = NULL,
  expr.data.file = NULL,
  tissue,
  genes,
  which.one = "rnaseq",
  silent = F,
  keep.grey = F,
  identicalNames = T,
  alt.gene.index = NULL,
  dupAware = T
)
}
\arguments{
\item{net}{The network can be either and object or a file}

\item{expr.data.file}{Again an object or a file, genes at columns, samples at rows.
It is expected that the genes are in the same order as they are expressed in the
network net$moduleColors vector. Use this parameter together with the net one and
which.one set to "new" for a network which is not in the DDBB.}

\item{tissue}{A tissue as it can be find in the network DDBB}

\item{genes}{The gene names. If set to NULL, the MM of all genes in the network is obtained}

\item{which.one}{The category the network belongs to. If not in the DDBB just ignore it}

\item{silent}{Set to true if you want no log}

\item{keep.grey}{Use grey genes too}

\item{alt.gene.index}{You can pass genes in a different order, use this index order for that}
}
\description{
Title Obtain the Module Membership of genes
This function works by getting the Pearson correlation between the
eigengene of the module and the gene(s).
}
