% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdb.R
\name{getNetworkFromTissue}
\alias{getNetworkFromTissue}
\title{getNetworkFromTissue - Get predefined network or create yours}
\usage{
getNetworkFromTissue(tissue = "SNIG", which.one = "exonic",
  only.file = F, genes = NULL, modules)
}
\arguments{
\item{modules}{}
}
\description{
getNetworkFromTissue - Get predefined network or create yours
}
