% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotMDS}
\alias{plotMDS}
\title{plotMDS - Testing gene sepparability with multi-dimensional
scalling.}
\usage{
plotMDS(rpkms.net, path, covs, covvars, label, n.mds = -1)
}
\arguments{
\item{rpkms.net}{A dataframe with the expression data with samples in columns.
Columns are assumed to use sample IDs as column names. These IDs should appear
as row names in the covariate file to properly index each value}

\item{path}{Folder to write the pdf plot generated}

\item{covs}{A data frame with all covariates for the samples. The rows must
be named with the sample IDs.}

\item{covvars}{The variables from the covs parameter to use in the plot}

\item{label}{A string for using at the plot, informative purposes}

\item{n.mds}{Number of genes to use, randomly sampled}
}
\value{

}
\description{
plotMDS - Testing gene sepparability with multi-dimensional
scalling.
}
