% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readTOM}
\alias{readTOM}
\title{Title Managing big TOM matrices}
\usage{
readTOM(filepref)
}
\arguments{
\item{filepref}{The prefix we used to previously save the matrix}
}
\value{
A squared TOM matrix disposed in the same order as genes in the
expression data matrix
}
\description{
Title Managing big TOM matrices
}
