% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneannot.R
\name{reportOnGeneIDs}
\alias{reportOnGeneIDs}
\title{Get known genes in a network}
\usage{
reportOnGeneIDs(tissue, which.one, genes)
}
\arguments{
\item{tissue}{c}

\item{which.one}{The category the network belongs to}

\item{genes}{The list of gene IDs you want to use. They can be as gene symbols or Ensembl IDs}
}
\value{
A vector of logical values
}
\description{
This method will return a vector of logical values with the same length as `genes`.
It will get the IDs of tne network in `tissue` and `which.one` convert both genes and ids
to Ensembl and check for matches. TRUE means the gene can be used in the analysis with
that network. FALSE means either the gene is not used in that network or that it could be
there with a different name or id format.
}
\examples{

}
