% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneannot.R
\name{reportOnGeneIDsGlobal}
\alias{reportOnGeneIDsGlobal}
\title{Get known genes in a set of networks}
\usage{
reportOnGeneIDsGlobal(tissues, categories, ids)
}
\arguments{
\item{tissues}{The networks as they appear in the network database}

\item{which.one}{The categories the networks belong to}

\item{genes}{The list of gene IDs you want to use. They can be as gene symbols or Ensembl IDs}
}
\value{
A matrix of logical values (genes in columns and networks in rows)
}
\description{
This method will return a matrix of logical values with the same columns as number of ids in `genes`.
For each network, it will call `reportOnGenIDs()`` as many times as different networks.
}
