% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{saveTOM}
\alias{saveTOM}
\title{Title Managing big TOM matrices
This function allows saving a bit TOM matrix aquared matrix, keeping only those
cells referring to genes clustering together in the same module}
\usage{
saveTOM(tom, clusters, filepref)
}
\arguments{
\item{tom}{The TOM matrix itself}

\item{clusters}{A vector with the labels (in the same order as in the expression data)
showing how genes cluster together}

\item{filepref}{This full path including file prefix will be used for the TOM files to
be saved}
}
\value{
No value returned
}
\description{
Title Managing big TOM matrices
This function allows saving a bit TOM matrix aquared matrix, keeping only those
cells referring to genes clustering together in the same module
}
