% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{testGeneSet}
\alias{testGeneSet}
\title{Title}
\usage{
testGeneSet(
  n.module,
  n.module.and.specific,
  total.specific,
  total.net,
  test = "fisher",
  oldform = F
)
}
\arguments{
\item{oldform}{}
}
\value{

}
\description{
Title
}
